Minor fixes to Snake (2020/ferguson1)/hint.text
Updated index.html as well but not through a discount program. This was
to make it easier not having to worry about adding the correct footer
since there aren't a lot of textual changes and I'm still not clear what
the Judges have set up with GitHub or what is even possible wrt that
part.