Add .gitignore to 2020/ferguson1 (Snake)
In particular it has globs for scripts that are copied by the Makefile
as well as alternative binaries (plus supplementary / helper binaries).

I'm aware that prog and alt.prog are not added to a root .gitignore but
as I said in comment
https://github.com/ioccc-src/winner/discussions/18#discussioncomment-3781474
I don't want to add to a root .gitignore file as that would not deal
with just my entry. Still it might want to be done. I've not done it for
my other entries and I won't in this pull request because the branch is
'snake-fixes'.