Rename *.text *.md
As luck would have it the only entries where this applied were my own.
The links/references to these files have been changed in the files.

An unfortunate problem is that in 2020/ferguson1 there are several
*.markdown files. I will have to go through these and figure out which
ones should be changed to .md and which ones can be removed. This can
happen later but in the meantime the references are to .md. I believe
the only difference is that the .md has the footer added by the judges.
This was a workaround with the fact that I wasn't sure (and still am not
entirely sure) the generation of the html files will work. It might be
that the .markdown files have the footer and the .md files do not but
this can be dealt with later anyway.