Minor fix in years-spoiler.html
There was a reference to Leo's fix to 1992/albert but although the
source code exists by the name given the email message did not. It
referred to 1992/albert/albert.fix.hint which did not exist. Previously
the README.md was actually that meant that the real entry README.md was
not generated into a html file. Thus I renamed the two files but only
now noticed that it's referenced in years-spoiler.html. I've renamed the
albert.bug.md to be albert.fix.hint.md to match the years-spoiler.html
wording (certainly another form could be contrived) but make it a
markdown file (even though it's not one).