Slight rewording in README.md
This change is one that is probably subjective which is why the
explanation below. It said:

    Please do NOT bookmark nor link to this experimental web site:

Now it's true that following 'not' a nor can be used before the second
or further of two[0] or more alternatives (first one being introduced in
this case by nor) to suggest that they each are false or do not happen
but somehow the wording:

    Please do NOT bookmark or link to this experimental web site:

flows better in my head. Again this is subjectively better and it might
even be a controversial thing but this feels better anyway.

Not that this is a double negative but on the subject of language a
couple fun facts about double negatives: in standard Spanish and Polish
they are used AND it was not until after the 16th century that they were
frowned up on in English.

Question: if !! (not not) cancels out the ! (meaning it's not negative
or rather it's true) is !!! (not not not or false) a not (false), a
double not (true) or a triple not (false) ? :-)

[0] Actually in this case it should be one or more alternatives but this
is either a mistake in OED or me being too literal (which is very
possible). 'Possibility' would be a better word to use as would 'choice'
(but there are certainly other, uh, choices).