Minor fix in 7154f9
I missed a '.text' file (chocolate-cake.text) in 2020/ferguson2 in
years.html.

As an aside on that famous chocolate cake recipe: since I included that
recipe in all my entries of 2020 and since the Enigma had a fun
challenge for that recipe: unless you want to play with the Enigma
machine and in so doing reward yourself with a wonderful double-layered
chocolate fudge cake recipe (the fudge being the icing) which is a
speciality of my mum's it would be better to just check the
ferguson1/chocolate-cake.html version as that's not enciphered. The
html versions are the recommended files to check at least with a
browser.