Fix 1987/biggar/README.md
There were two instructions for building but the one (with the compiler
line) did not work any more and once fixed it appears to be the same as
what the Makefile does so I've removed that example.

As well I fixed the file to refer to 'biggar' and not 'markb'.