Remove 1986/stein/stein.orig.c
Since the original program does not compile and more than one of us
fixed it (the judges and me and probably others as well) and the fix is
in stein.c we remove the stein.orig.c.

Updated Makefile so this would not be a problem. Updated README.md to
not refer to make alt.