Get most 1993 entries to compile
With the exception of lmfjyh which relies on a bug in gcc (that was
fixed in gcc 2.3.3) we have got all 1993 entries to compile under linux.
A couple will not compile under macOS and this is because gcc under
macOS is actually clang despite what many might have you believe (see
below) which is more strict about the second and third arguments of
main().

In a couple entries we used patches from Yusuke Endoh. Thank you Yusuke!
In others we (that is I aka Cody / @xexyl) fixed them (myself). There
was a question to be answered about whether to add #include lines to the
files themselves or add -includes to the Makefile. Since there are
pending changes to the Makefiles which will affect the CFLAGS (and
there's no CINCLUDE variable in these Makefiles yet) we opted for
updating the source files themselves.

One more note about the Makefiles: the 1993/Makefile was modified but I
added an XXX note to it to indicate that it should be fixed once the
other Makefiles are fixed. Specifically it now allows from 1993 one
to run `make' and compile the entries. The comment has what it used to
read and should be but for now is not because of a problem in the other
Makefiles.

As far as gcc under macOS as you can see it is not actually gcc but
clang:

    $ /usr/bin/gcc -v 2>&1|head -n 1
    Apple clang version 14.0.0 (clang-1400.0.29.202)

If you were to use clang -v under linux you might see something like:

    $ clang -v 2>&1|head -n 1
    clang version 15.0.7 (Fedora 15.0.7-1.fc37)

and with gcc:

    $ gcc --version |head -n 1
    gcc (GCC) 12.2.1 20221121 (Red Hat 12.2.1-4)

If you do by chance have gcc compiled and installed under macOS these
entries will probably work there.