Minor fixes in 2020/kurdyukov4
The Makefile referenced a URL that no longer is valid (or less valid?).

The README.md referred to shakespeare.txt but the file (extracted from
the file downloaded from the curl invocation in the Makefile) is
Shakespeare.txt (in some places it was capitalised).

There were some formatting fixes in the README.md file too.

There is a .gitignore added but not in this commit - an upcoming commit.
It has the two files in that gzip file added but since it also had other
files I've kept it for a separate commit. A question of whether the
files should just be part of the entry is another to maybe consider but
for now it's not.