Fixes in 2020/kurdyukov2
The Makefile now chmod +x makegif.sh as it's referred to in the
README.md.

Change word 'Depends' to 'Dependencies' in makegif.sh to read a bit
easier.

Make README.md a bit easier to read (as English is not Ilya's first
language).

Made the command Ilya suggests to make GIFs so that it makes use of the
command the judges suggested, noting that it's just an example.

Added middle name to my name to match my name as a winner.

Of course if Ilya objects to any of this (or any other entry of his) I'm
certainly okay with rolling back the changes (same with his other
entries) but I believe that this will make it a bit easier to try out
the entry.