Typo fixes in 2020/yang/README.md
It appears that the entry name is 'Mile' but since it's compiled to be
'prog' the command invocations should refer to ./prog not ./mile. I kept
the other references to Mile as is.

I'm aware that this entry does not seem to work under macOS but it could
be that I've been working on this year too much and I need to finish up
(fortunately I have only one more entry to look at). But it might be
that it's simply not functional under macOS.