Remove */rules.html
Yes this is on the guidelines branch but I forgot about the html files
(briefly) when on that branch and I don't want to go back to the other
branch and then checkout another branch that does not have these
changes. Maybe they should have both been on the same branch but I want
to avoid problems with conflicts (even though these should be fine).