Rename 1985/lycklama.orig.c -> 1985/lycklama.alt.c
Updated README.md to refer to the alt version instead. Note that it's
there because it provides fun output and also shows how it originally
looked (which is quite different from the fixed [for modern compilers]
version that I believe I made in an earlier commit but if so I'm not
sure which one).