Typo fixes in 1992/nathan
In nathan.c there was also a typo fix but since it's not the original
entry (or any thing like it) it's okay to do this. This is actually the
subject of a discussion to be had on GitHub which I'll be addressing in
a bit.

In README.md there was also a typo fixed.