Get all 1988 entries to compile
For 1988/dale it will not compile under macOS. I was able to get it to
almost compile under macOS but there was one problem. As well this 'fix'
will make it not compile under linux. One change was required which was
changing the #define _ define followed by #_ foo to be just '#define
foo' etc.

Added to 1988/isaak/.gitignore isaak.output.

1988/spinellis won't compile with clang.

1988/phillipps will compile but it is likely to segfault at least in
some systems (maybe not all).

Updated many README.md files for the changes above (and probably other
changes I've neglected to write in this log). Most changes in the readme
files are probably just formatting and probably typos as well as how to
compile and run entries nowadays. There might have been other changes.