Address 2019 entries
In at least one case some -D flags had to be added to the Makefile.

Typo fixes and formatting fixes in some README.md files.

For the simply amazing 2019/karns I added a demo.sh which does a cat on
the two files (prog.c and maze) and then sleeps for two seconds then
runs the program then sleeps for two seconds then resets the terminal
(so the colours are messed up). This was added to the README.md file.

Several .gitignore files were added.

I'm aware that the graphics editing one does not now work under macOS
but I'm not sure if this is all versions so I just noted that it doesn't
seem to work with recent versions of macOS as I don't have a way to test
before Ventura (the current version).