Get 1986/marshall to compile and work with clang
Choo choo!

This did not take much that work but it required some casts because the
third arg to main() was an int and now it's a char **. I tried to keep
the layout as close to the original. The line lengths are the same as
the original but some spaces are now non-spaces. I noted this in the
README.md.

As well in the README.md I have added an author's comments section.
Still working on other README.md files for this. However I might try and
get some entries to compile that previously did not. There was another
entry in 1986 that this was not so simple to do so I moved on. Maybe I
will revisit that another time (1986/holloway).