Move 2014/vik bug notice from years to README.md
As the years.html file will be generated with a tool (yet to be written)
having the bug notice will (greatly)? complicate the generation. Thus
the bug notice (which I discovered when working on 2020/ferguson2) was
moved to 2014/vik/README.md.

But why? Because it is a good idea, I think, to note that although
mostly correct, there is a minor problem with it. I don't have a fix for
it but whether I ever look at that I do not know.

I also made a minor change in 2020/ferguson2/README.md: changed a full
stop to an exclamation mark as it seemed better.