Get 1987/westley to compile
Unfortunately this ruins the symmetry but I did make it so that the
problematic code is still shown just commented out. The problem was one
can no longer assign to a cast.

The original version was preserved in westley.alt.c, a worthy thing I
think to do as symmetry is magical and it should be shown how it once
worked without having to look at the archive.