fix Makefile and README.md for 2000/natori
There was an attempt to get natori.c to compile.
We moved the 2 includes to the compile line
restoring the code as a 1-liner.

We forced the types of args iun main to conform to
bogons introduced by the ANSI-C committee.

Nevertheless, the code still fails to compile:

    assignment to cast is illegal, lvalue casts are not supported

This will be a TODO fix for later.