Get 1989/westley to compile with clang
Some important notes here should be made. First of all with clang the
recommended commands in the README.md (which was also formatted and typo
fixed) will seemingly generate the correct files but still only
westley.c and ver0.c will compile with clang. With gcc they do compile
and seem to work as expected.

However I found a bug in the original code too where ver2 will sometimes
segfault. This also goes for the fix (again when compiled with gcc).

I did get ver[123].c to also compile but this seemed to cause invalid
output in some of the commands. That being said it might (though I don't
think so at this time) have been not knowing exactly what should be
displayed. Maybe I will look at it in the future but at least it now
works with clang even if not all of the generated files will compile
that way. Still macOS users can now take a look at the different output
of the program which seems good to me (it does not even have to be
source code that is output).

These things have been noted in the README.md as well.