Fix 1999/brnstnd
This was a fun one to fix. There were two invalid operands to binary
expression errors to fix which was resolved with the comma operator.
More fun was dealing with a C pre-processor change. The macro C was used
like e/C to do e/=G but this is no longer allowed. The solution was to
not use the C macro and instead just do the /= etc.