README.md and Makefile files from various years
These are mostly just formatting fixes but there might be some typos
here and there.

These years have not been checked thoroughly but I made fixes and want
them in for when I do work on them.

As far as the Makefiles go: notes were added that correspond to the
notes in the README.md files which is why I have put them in the same
commit.