Further fix 1996/dalbec
I made the recommended change by the author so that each number is on a
line by itself rather than one long line. I changed the try section to
stop at the output that was the output noted by the judges. This was
possible with the change of making it a newline rather than a space
after each number.

I had originally proposed the change in commit
9b6b5a0376f3f443d8a8ef0721ae9418d06fdc61 but I did not implement it
because the output of the command versus what the try section included
was misleading (maybe that was correct :-) ) and when looking at Yusuke
Endoh's output I missed that it continued after the number (he had '...'
but I missed it in the long line of output). I linked to his analysis as
well. Thank you Yusuke for confirming it is correct output!