Fix segfault in tbr.alt.c, swap tbr.alt.c code
The segfault has been fixed when EOF is sent to the program (causing
fgets() to return NULL).

Since fgets() is far safer than gets() and since the gets() version will
intersperse the warning and the first line of output I have changed the
alt version to be the original (fixed to compile with modern compilers)
and the tbr is the one with fgets().