Further fix 2004/newbern
Landon and I both individually fixed it to work with clang. It appears
that both work just as well but I made it so that the argv (now ll, was
l) in main() is not the name of a function that is defined earlier in
the file. Although Landon's fix worked this felt like a better idea
especially as it seems like the argv he made was not referenced in
main() but it should be.