Get 2006/monge to link
The problem was that SDL2 removed two functions that are used in this
code. Since SDL1 is still available and since changing it to SDL2 is
more complicated than it is worth I've decided to keep it as SDL1,
changing sdl2-config to sdl-config. Nevertheless it actually segfaults
or it does at least in macOS. I have not tested it in linux but I will
in time. At this point I will try debugging it but at least now it will
link.

Typo fixes in README.md and update the reference to SDL2 to be SDL1. I
noted that currently it does not work at least under macOS. I've not yet
put it in the bugs.md file but will in time if I don't get it to work
(again once I test it under linux).