2006/monge requires x86/x86_64 CPUs
The segfault I noted in c40a061a2d114fd1e5c1549c8f9e1ab0421adfed appears
to be because it requires x86/x86_64 architecture and I tested it on the
arm64 (MacBook Pro M1). I will still in time test it under linux to make
sure it works but that's for another time.

With this in mind the entry should at least theoretically work at least
if you have an x86/x86_64 CPU.