Update bugs.md
The segfault in 1996/westley I fixed some days ago so there's no need
for it to be in bugs.md.

Added the requirement for 32-bit arch and no optimiser in 2005/giljade
but suggesting it might be possible to fix it for 64-bit systems.
Probably this should be in a different section but I put it in the known
bugs to fix (even though it's not necessarily a bug).

Typo fix in 1990/jaw.

To see 1984/mullender in action we don't need an emulator: one can use
2015/endoh3 instead. I noted this, linking to it, but also noted that
otherwise one would need an emulator for the entry.

2004/gavin actually works. The problem is it doesn't work under macOS
at least with the M1+ chip (might or might not work with Intel). I still
kept the recent changes to the file but I noted it's not related to the
problem; it's an architecture issue.