Minor updates to 2005/mynx
Though this will not work as is I have changed the code to scan for
https as well (it exited before). This is so if someone comes up with a
clever command line to make it work it can. A secure connection has to
be established before http commands can be given but now it at least can
connect (though it's been a while since I've played with sockets so
maybe I'm forgetting some detail - one does have to specify the port but
I gave an example in the README.md file for that).

In the README.md I pointed out that one has to specify http:// (or else
https though with the caveat noted above). I also explained how to
specify an alternative port.

I also noted the https issue in the bugs.md file in the INABIAF section
('it's not a bug it's a feature').

As well I changed the try command in the entry README.md file to refer
to a page that does not use https as trying it on the ioccc website
(what it used to) will just return a 301 error.