Fix 1993/vanb for clang
The problem was the usual suspect where any arg other than the first was
not a char **. Instead O5 is no longer defined to be main() and it is
its own function now which main() calls. The third arg starts out as a
global variable that is passed into the function (which is recursive).

Now the output is correct.