Update 1998/schweikh3/README + add dupe file
The formatting of README.md has been changed.

Added a try section and a duplicate file to the directory (copy of
samefile.1) so that the users can see how it looks. The copy is
samefile1.1.dupe. Yes dupe means to deceive and also a victim of being
deception but it also means duplicate. For the IOCCC though it actually
makes sense if a file (in fact all the source files :-) and thus you're
a dupe! :-)