Fix 2006/sloane for clang
The problem was that the second arg was not of type char ** but an
implicit int and broken record though I am clang forces the types to be
correct without any option to disable this.

As it happens though the second arg is not even used so I removed it as
clang does allow this for some reason. This helps with keeping the
original format as close as possible. This triggers a different warning
but that's okay.

I also added a warning to the README.md about how it is a lot of change
in text very quickly as this could cause some problems for some people.