Fixes and improvements to bugs.md
A line for one entry was in the wrong entry.

I added that I'm working on them (with some humour injected there) but
others may try too .. if they are okay making people very sad :-)

Added more information / examples in some of the status sections and
possibly some of the entries that are listed.