Fix 1998/schnitzi for linux
It segfaulted due to wrong type of a variable.

It still crashes under macOS which I will be looking at in the near
future but it at least works for linux.

I noted in the bugs.md a compiler warning that should not be changed as
it's not a bug and in fact 'fixing' it will introduce a bug.