Fix German vier in 2012/kang
Although the judges do note that the entry only gets some of the
languages that it supports right it seems to me it should accept vier
not fier esp as it gets other numbers right. It did not appear to be a
situation where the entry predicts the misspelt words (trying vier did
not work) so I think this is correct. If this was intended I am
certainly fine with it not being fixed but it appears to be a bug to me.
I did a variety of other tests to make sure other numbers are still
okay.

An example where it detects misspelt words:

    $ echo secht | ./kang
    6

... instead of sechs for six. But it also gets sechs right. It does not
get zehn right but it does get eins right. But as the judges say this is
okay.