2018/ferguson fixes and minor updates
Fix test.sh for shellcheck.

Minor fixes to README.md including moving the author information (that
is my information :-) ) to the first column of the file as not having it
prevents links from being generated right. This is a problem for the
other README.md files too and I'll be noting it on GitHub.

Some humour with html <del>..</del>.

Modified portability section to make it a bit simpler and suggest that I
won't be updating the list. I only include the original systems tested
for historical purposes. I won't be updating the list over time as that
would be annoying to do and would just make the already very long
document needlessly longer. It's already been almost 5 years since it
was updated.

Add a string to test-strings.txt.

Warning: these are most likely not the final fixes or updates for my
entries as Landon has pointed out :-) but there shouldn't be many more
for now or hopefully for a long time. Unless I see something I won't
feel the need to fix anything.

Update 2018/burton2/README.md to have my middle name.