Update several of the files from 2015
With all README.md files I changed the copyright years to refer not to
2016 but 2015 as the year was 2015 and 2016 did not even have a contest.

I fixed formatting and some typos in a few of the README.md files but
mostly the changes are the copyright years.

As a reminder to myself I have only looked at (counting those of commit
fa31f80b39fb310024695074f92d229389806f4e) burton, dogon, duble and part
of endoh1. The rest will come in a future commit or maybe commits.