Fix 2015/endoh3/README.md
Fix dead link. Apout is no longer available at the location in the
README.md. Instead it's on GitHub. I checked it and it's the same
author, Warren Toomey of TUHS so I know it's good (or at least the
program is: not sure if the repo is owned by him).

Format fixes.

Fix ioccc.org link (though it shouldn't be necessary since mullender.c
is actually in the repo).