Add 1989/robison to bugs.md
The question of whether this is worth calling a bug (or a bug that
should be fixed) or not is certainly one to consider and I put that in
there. Nevertheless with non-binary digits the program will crash.
Should this be fixed? I don't know. I'll certainly look at it in time
though.