Get 1992/lush to work with gcc - won't with clang
I used a patch from Yusuke Endoh to get it to work with gcc. Due to the
different messages with different compilers it's expected that this
won't work with every compiler.

I provided a 'runme.sh' script to run the command in the correct way.

I made the entry use fgets() not gets().

I updated bugs.md pointing out that it does not work with clang and
alternative code is welcome.