Add faq.md, remove faq.html
I tried to keep the formatting as close to the original as possible.
It's very likely links have to be fixed and it's very likely formatting
has to be fixed.

Some things necessarily had to change but very few did in fact need to
change.

As tempting as it was to change it to British English I did not (though
I started to when fixing a link but I caught myself in time) :-)

Naturally as this is a major change problems are expected but these will
be worked out. I guess the next step would be to do this for other raw
html files but I don't fancy doing more than one in a day. This can be
discussed later though as some will be automatically generated anyway so
no need for a markdown file (faq.md was discussed as a good idea).