Improve 1984/mullender alt code
Now you can specify the number of microseconds to sleep in between
writes! This makes it easier to experiment with different speeds as well
as solve any problems if your CPU is too slow or too fast

BTW: is there such a thing as too fast a CPU ? :-) actually yes for
certain code which is probably not as uncommon as you think :-) ).