Add author comments to 1984/anonymous
Some years after the entry was published the judges asked them if they
wished to remain anonymous. They said yes but added that they are well
known for their there connection to the C language.

I fixed at the same time the typo that was in this quote as it seemed
that it's better to be clear than to keep a quote exactly as is
especially as it might have not been a copy paste in the first place.