Update 2018/giles + faq.md
Mostly this update is about SDL and SDL2 and how to compile and run
entries that make use of it. But I first saw reference to this idea in
2018/giles/README.md.

I added to macOS section MacPorts which is another very popular package
manager for macOS (I personally don't like Homebrew and I use MacPorts
instead).

Minor fixes in 2018/giles/README.md (formatting plus make a URL an
actual link).