Fixes in 2018/ciura/README.md + Makefile
Formatting fixes and possibly typos (but mostly formatting).

On bold code blocks: I removed the bold because it appeared that some of
the formatting was messed up. I think this is okay though rather than
trying to fix the formatting of bold code blocks.

Set alt related target variables in Makefile so that make alt actually
works.