Fixes in 1994/westley
Format fixes in README.md.

Added note from author about terminal configuration from Brian at the
top of the README.md and the Makefile so that it's easier to see.

Using the spoiler section from Brian I converted it to compiler commands
(in the README.md) and made a spoiler.sh script to run them - to make it
easier to see the game in full for those who just want to see it and not
try it.